Split facet-core/types into multiple modules, prepare for Arc<T> etc.
